regression/requirements.txt
matplotlib==3.7.4
munch==2.5.0
numpy==1.24.4
PyYAML==6.0
scipy==1.10.1
seaborn==0.13.0
torch==2.1.0+cu121